chore(rollup): Hoist entrypoint index.js files
<NAME>/index.js files are now hoisted up one folder to become <NAME>.js
files.

node_modules imports are now all bundled inside _vendor/vendor.js.

The rewrite-filenames plugin has been moved to a separate file.

Handle circular xmlbuilder and jzip targets.
